+++ src/routes/index.tsx
@@ -1,7 +1,6 @@
 import { createBrowserRouter, Outlet } from 'react-router-dom';
 
 import Header from '@/components/Common/Header/Header';
-import PrivateRoute from '@/components/PrivateRoute';
 import PSHeader from '@/components/PSHeader/PSHeader';
 import {
   CreateRoomPage,
@@ -17,6 +16,7 @@ import {
 
 import { CheckUserLogin } from './CheckUserLogin';
 import { PATH } from './path';
+import PrivateRoute from './PrivateRoute';
 
 export const router = createBrowserRouter([
   {
